+++ internal/testdata/circular_fk_models/models.go
@@ -1,0 +1,13 @@
+package circular_fk_models
+
+type Location struct {
+	LocationID string `gorm:"primaryKey;column:locationId;"`
+	EventID    string `gorm:"uniqueIndex;column:eventId;"`
+	Event      *Event `gorm:"foreignKey:locationId;references:locationId;OnUpdate:CASCADE,OnDelete:CASCADE"`
+}
+
+type Event struct {
+	EventID    string    `gorm:"primaryKey;column:eventId;"`
+	LocationID string    `gorm:"uniqueIndex;column:locationId;"`
+	Location   *Location `gorm:"foreignKey:eventId;references:eventId;OnUpdate:CASCADE,OnDelete:CASCADE"`
+}
