+++ internal/testdata/models/models.go
@@ -19,15 +19,3 @@ type Toy struct {
 	ID   uint
 	Name string
 }
-
-type Location struct {
-	LocationID string `gorm:"primaryKey;column:locationId;"`
-	EventID    string `gorm:"uniqueIndex;column:eventId;"`
-	Event      *Event `gorm:"foreignKey:locationId;references:locationId;OnUpdate:CASCADE,OnDelete:CASCADE"`
-}
-
-type Event struct {
-	EventID    string    `gorm:"primaryKey;column:eventId;"`
-	LocationID string    `gorm:"uniqueIndex;column:locationId;"`
-	Location   *Location `gorm:"foreignKey:eventId;references:eventId;OnUpdate:CASCADE,OnDelete:CASCADE"`
-}
